Add files via upload
This file contains the sales record of a company from 2010 to 2017. A dashboard using the Sales Records file while ensuring data driven insights are present in the dashboard.
POWER QUERY (M-code) and POWER BI (DAX LANGUAGE) were used in this projects to get insights and make data driven decision from the dashboard attached in the read me .file. Also, the raw data is attached in the file above.